% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{as.list.capture_diff}
\alias{as.list.capture_diff}
\title{Convert capture_diff to list}
\usage{
\method{as.list}{capture_diff}(x, ...)
}
\arguments{
\item{x}{A `capture_diff` object.}

\item{...}{Ignored.}
}
\value{
A named list of field comparisons.
}
\description{
Convert capture_diff to list
}
