% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.list.capture_env}
\alias{as.list.capture_env}
\title{Convert capture_env to list}
\usage{
\method{as.list}{capture_env}(x, ...)
}
\arguments{
\item{x}{A `capture_env` object.}

\item{...}{Ignored.}
}
\value{
Named list of contained objects.
}
\description{
Convert capture_env to list
}
