% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.list.capture_info}
\alias{as.list.capture_info}
\title{Convert a capture_info object to a list}
\usage{
\method{as.list}{capture_info}(x, ...)
}
\arguments{
\item{x}{A `capture_info` object.}

\item{...}{Ignored.}
}
\value{
Named list of metadata.
}
\description{
Convert a capture_info object to a list
}
