% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{attach_capture_to_objects}
\alias{attach_capture_to_objects}
\title{Attach capture metadata to objects}
\usage{
attach_capture_to_objects(objects, capture)
}
\arguments{
\item{objects}{Named list of objects.}

\item{capture}{A `capture_info` object.}
}
\value{
List of objects with capture attribute set.
}
\description{
Attach capture metadata to objects
}
\keyword{internal}
