% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{build_capture_env}
\alias{build_capture_env}
\title{Capture and attach metadata to multiple objects}
\usage{
build_capture_env(
  object_names,
  code_lines,
  env,
  script_path = NULL,
  tags = NULL,
  digest_algo = "sha256",
  verbose = FALSE,
  include_session = FALSE
)
}
\arguments{
\item{object_names}{Character vector of object names.}

\item{code_lines}{Character vector of code lines.}

\item{env}{Environment in which to locate objects.}

\item{script_path}{Optional script path.}

\item{tags}{Optional character vector of tags.}

\item{digest_algo}{Digest algorithm to use.}

\item{verbose}{Logical; whether to print capture summary.}

\item{include_session}{Logical; Should session info be included in metadata}
}
\value{
A `capture_env` object.
}
\description{
Capture and attach metadata to multiple objects
}
\keyword{internal}
