% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{build_capture_info}
\alias{build_capture_info}
\title{Build capture_info metadata object}
\usage{
build_capture_info(
  object_name,
  timestamp,
  script_path,
  code_lines,
  dependencies,
  script_checksum,
  seed,
  session,
  object_digest,
  git_info,
  tags,
  package_versions,
  warnings,
  digest_algorithm
)
}
\arguments{
\item{object_name}{Name of the object.}

\item{timestamp}{POSIXct timestamp of capture.}

\item{script_path}{Path to the originating script.}

\item{code_lines}{Code lines used to generate the object.}

\item{dependencies}{List of ScriptNodeInfo objects.}

\item{script_checksum}{Checksum of the script file.}

\item{seed}{set.seed() line found in code.}

\item{session}{Output of sessionInfo().}

\item{object_digest}{Digest hash of the object.}

\item{git_info}{Git metadata list.}

\item{tags}{Optional tags.}

\item{package_versions}{Named character vector of package versions.}

\item{warnings}{List of warning strings.}

\item{digest_algorithm}{Digest algorithm used.}
}
\value{
A `capture_info` object.
}
\description{
Build capture_info metadata object
}
\keyword{internal}
