% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{package}
\name{captuR}
\alias{captuR-package}
\alias{captuR}
\title{R Object Capture and Context Tracking}
\description{
Save R objects with associated context: code, dependencies, digests, and Git/script metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-a-schultz/captuR}
  \item Report bugs at \url{https://github.com/m-a-schultz/captuR/issues}
}

}
\author{
\strong{Maintainer}: Michael Schultz \email{m-a-schultz@protonmail.com}

}
\keyword{internal}
