% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{capture_checksum_valid}
\alias{capture_checksum_valid}
\title{Validate a script checksum against the capture metadata.}
\usage{
capture_checksum_valid(object)
}
\arguments{
\item{object}{An object with a `capture` attribute.}
}
\value{
TRUE if the script is unchanged, FALSE otherwise.
}
\description{
Validate a script checksum against the capture metadata.
}
