% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{capture_diff}
\alias{capture_diff}
\title{Compare capture metadata for two objects}
\usage{
capture_diff(object1, object2, ignore = NULL)
}
\arguments{
\item{object1}{First captured object or capture_info}

\item{object2}{Second captured object or capture_info}

\item{ignore}{Optional character vector of field names to ignore}
}
\value{
An object of class `capture_diff`
}
\description{
Compare capture metadata for two objects
}
