% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{capture_digest}
\alias{capture_digest}
\title{Get digest from capture metadata}
\usage{
capture_digest(object)
}
\arguments{
\item{object}{An R object with capture metadata.}
}
\value{
Digest string or NULL if not available.
}
\description{
Get digest from capture metadata
}
