% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_helpers.R
\name{capture_environment_snapshot}
\alias{capture_environment_snapshot}
\title{Snapshot objects from environment}
\usage{
capture_environment_snapshot(object_names, env = parent.frame())
}
\arguments{
\item{object_names}{Character vector of object names}

\item{env}{Environment to read from}
}
\value{
Named list of object values
}
\description{
Snapshot objects from environment
}
\keyword{internal}
