% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture_object}
\alias{capture_object}
\title{Attach metadata to a single object}
\usage{
capture_object(
  object,
  script_path = NULL,
  tags = NULL,
  digest_algo = "sha256",
  verbose_guess = FALSE,
  contextual_code = TRUE,
  include_session = FALSE
)
}
\arguments{
\item{object}{An R object to capture.}

\item{script_path}{Optional path to source script.}

\item{tags}{Optional tags.}

\item{digest_algo}{Digest algorithm to use.}

\item{verbose_guess}{Logical; whether to print guess path warnings.}

\item{contextual_code}{Logical; if TRUE, return all code from set.seed() to object creation.}

\item{include_session}{Logical; Should session info be included in metadata}
}
\value{
The input object with an attached `"capture"` attribute of class
  `capture_info`, which encapsulates all associated metadata for reproducibility
  and tracking.
}
\description{
Analyzes the provenance of an R object and attaches metadata to it, without
saving it to disk. Metadata includes the code that generated the object, its
digest, seed, dependencies, and script/Git information.
}
\seealso{
[capture_info()], [get_capture()]
}
