% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{capture_script_path}
\alias{capture_script_path}
\title{Get the script path from capture metadata}
\usage{
capture_script_path(object)
}
\arguments{
\item{object}{A `capture_info` object.}
}
\value{
Character string of the script path.
}
\description{
Get the script path from capture metadata
}
