% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{capture_summary}
\alias{capture_summary}
\title{Capture Metadata Summary for Multiple Objects}
\usage{
capture_summary(objects)
}
\arguments{
\item{objects}{A named list of R objects (or captured environment).}
}
\value{
A data.frame with metadata summary.
}
\description{
Returns a data frame summarizing key metadata from a list of captured objects.
}
