% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{capture_tags}
\alias{capture_tags}
\title{Get tags from capture metadata}
\usage{
capture_tags(object)
}
\arguments{
\item{object}{An R object with capture metadata.}
}
\value{
Character vector of tags or NULL if not available.
}
\description{
Get tags from capture metadata
}
