% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{capture_timestamp}
\alias{capture_timestamp}
\title{Retrieve the capture timestamp}
\usage{
capture_timestamp(object)
}
\arguments{
\item{object}{A captured object.}
}
\value{
POSIXct timestamp or NULL.
}
\description{
Retrieve the capture timestamp
}
