% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{check_capture}
\alias{check_capture}
\title{Check the consistency of the generating script with the recorded checksum.}
\usage{
check_capture(capture, verbose = TRUE)
}
\arguments{
\item{capture}{The capture metadata object (of class 'capture_info').}

\item{verbose}{Logical indicating whether to print messages (default: TRUE).}
}
\value{
TRUE if the script is consistent with the saved checksum.
}
\description{
Check the consistency of the generating script with the recorded checksum.
}
