% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{compare_fields}
\alias{compare_fields}
\title{Compare two values with optional ignore logic}
\usage{
compare_fields(x, y, keys, ignore = NULL)
}
\description{
Compare two values with optional ignore logic
}
\keyword{internal}
