% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{compare_outputs}
\alias{compare_outputs}
\title{Compare original and rerun outputs for a list of objects}
\usage{
compare_outputs(outputs, original, rerun)
}
\description{
Compare original and rerun outputs for a list of objects
}
\keyword{internal}
