% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{detect_created_objects}
\alias{detect_created_objects}
\title{Detect created objects from a dependency list}
\usage{
detect_created_objects(dependency_list)
}
\arguments{
\item{dependency_list}{List of ScriptNodeInfo objects}
}
\value{
Character vector of created object names
}
\description{
Detect created objects from a dependency list
}
\keyword{internal}
