% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{export_capture_code}
\alias{export_capture_code}
\title{Export generating code from captured object.}
\usage{
export_capture_code(object, file, header = TRUE)
}
\arguments{
\item{object}{A captured object.}

\item{file}{File path for output.}

\item{header}{Include header metadata? Default TRUE.}
}
\value{
File path (invisibly).
}
\description{
Export generating code from captured object.
}
