% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{extract_dependencies_from_code}
\alias{extract_dependencies_from_code}
\title{Extract dependencies from code lines}
\usage{
extract_dependencies_from_code(code_lines = NULL, script_path = NULL)
}
\arguments{
\item{code_lines}{Character vector of R code lines}

\item{script_path}{Path to the script file}
}
\value{
List of ScriptNodeInfo objects
}
\description{
Extract dependencies from code lines
}
\keyword{internal}
