% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{get_capture}
\alias{get_capture}
\title{Access capture metadata from an R object}
\usage{
get_capture(object)
}
\arguments{
\item{object}{An R object saved using `capture_save()`.}
}
\value{
A `capture_info` object or NULL if unavailable.
}
\description{
Retrieves the metadata stored during object capture.
}
