% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_object_dependencies}
\alias{get_object_dependencies}
\title{Get Code Dependencies for an Object}
\usage{
get_object_dependencies(obj_name, script_path = NULL)
}
\arguments{
\item{obj_name}{Character name of the object to trace.}

\item{script_path}{Path to the script file (optional). Will guess if NULL.}
}
\value{
A list of `ScriptNodeInfo` objects or NULL.
}
\description{
Uses `CodeDepends` to find the statements in a script that generated a specific object.
}
