% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_helpers.R
\name{get_package_versions}
\alias{get_package_versions}
\title{Get package versions from CodeDepends dependencies}
\usage{
get_package_versions(dependencies)
}
\arguments{
\item{dependencies}{List of ScriptNodeInfo objects}
}
\value{
Named character vector of package versions
}
\description{
Extracts used packages based on namespaced functions in CodeDepends analysis.
}
