% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_helpers.R
\name{get_script_metadata}
\alias{get_script_metadata}
\title{Get metadata from a script path}
\usage{
get_script_metadata(path, algo = "sha256")
}
\arguments{
\item{path}{File path to the R script}

\item{algo}{Digest algorithm to use}
}
\value{
Named list with script metadata
}
\description{
Get metadata from a script path
}
\keyword{internal}
