% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{has_capture}
\alias{has_capture}
\title{Check if an object has capture metadata}
\usage{
has_capture(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
TRUE if the object has a capture attribute of class `capture_info`, FALSE otherwise.
}
\description{
Determines whether the object has a valid capture metadata attribute.
}
