% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{identify_inputs}
\alias{identify_inputs}
\title{Identify input objects from a dependency list}
\usage{
identify_inputs(dependencies, obj_name = NULL)
}
\arguments{
\item{dependencies}{(Optional) name of a focal object}
}
\value{
Character vector of input object names
}
\description{
Identify input objects from a dependency list
}
\keyword{internal}
