% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{identify_outputs}
\alias{identify_outputs}
\title{Identify output objects from a dependency list}
\usage{
identify_outputs(dependencies)
}
\arguments{
\item{dependencies}{List of ScriptNodeInfo objects}
}
\value{
Character vector of created object names
}
\description{
Identify output objects from a dependency list
}
\keyword{internal}
