% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{list_capture_dependencies}
\alias{list_capture_dependencies}
\title{List input dependencies from capture metadata}
\usage{
list_capture_dependencies(object)
}
\arguments{
\item{object}{A captured object.}
}
\value{
A list of dependencies or NULL.
}
\description{
List input dependencies from capture metadata
}
