% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{print.capture_diff}
\alias{print.capture_diff}
\title{Print method for capture_diff}
\usage{
\method{print}{capture_diff}(x, ...)
}
\arguments{
\item{x}{A `capture_diff` object.}

\item{...}{Ignored.}
}
\description{
Print method for capture_diff
}
