% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.capture_env}
\alias{print.capture_env}
\title{Print method for capture_env}
\usage{
\method{print}{capture_env}(x, ...)
}
\arguments{
\item{x}{A `capture_env` object.}

\item{...}{Ignored.}
}
\value{
The object (invisible).
}
\description{
Displays a summary of the captured environment.
}
