% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.capture_info}
\alias{print.capture_info}
\title{Print summary of capture metadata}
\usage{
\method{print}{capture_info}(x, ...)
}
\arguments{
\item{x}{A `capture_info` object.}

\item{...}{Ignored.}
}
\description{
Print summary of capture metadata
}
