% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{[[.capture_env}
\alias{[[.capture_env}
\title{Extract an object from capture_env}
\usage{
\method{[[}{capture_env}(x, i)
}
\arguments{
\item{x}{A `capture_env` object.}

\item{i}{Index of object to extract.}
}
\value{
The extracted object.
}
\description{
Allows access to objects via `$` operator.
}
