% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{summary.capture_diff}
\alias{summary.capture_diff}
\title{Summary method for capture_diff}
\usage{
\method{summary}{capture_diff}(object, ...)
}
\arguments{
\item{object}{A `capture_diff` object.}

\item{...}{Ignored.}
}
\description{
Summary method for capture_diff
}
