% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{summary.capture_info}
\alias{summary.capture_info}
\title{Summary method for capture metadata}
\usage{
\method{summary}{capture_info}(object, ...)
}
\arguments{
\item{object}{A `capture_info` object.}

\item{...}{Ignored.}
}
\value{
A named list summarizing key metadata.
}
\description{
Summary method for capture metadata
}
