% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_funcs.R
\name{view_capture_code}
\alias{view_capture_code}
\title{View the generating code from capture metadata}
\usage{
view_capture_code(object)
}
\arguments{
\item{object}{A captured object.}
}
\description{
View the generating code from capture metadata
}
